@@ -1,120 +1,103 @@
 [build-system]
-requires = ["setuptools>=70.0.0"]
-build-backend = "setuptools.build_meta"
+requires = ["hatchling", "hatch-vcs"]
+build-backend = "hatchling.build"
 
 [project]
 name = "vm_clockify"
-description = "clockify api usage to upload into youtrack"
-keywords = []
+description = "clockify api usage to upload into youtrack."
+keywords = ["youtrack", "clockify", "time"]
 
-version = "0.1.0"
-requires-python = ">=3.12.0"
+version = "1.0.0"
+requires-python = ">=3.13.0"
 readme = "README.md"
 license = { text = "MIT" }
 
 authors = [{ name = "MVladislav" }]
 
 dependencies = [
-  "setuptools>=75.0.0",
-  "bs4==0.0.2",
-  "click==8.1.7",
-  "coloredlogs==15.0.1",
-  "holidays==0.35",
-  "httpx==1.0.0b0",
-  "pydantic==2.7.0",
-  "pydantic-settings==2.6.1",
-  "pytz==2023.3.post1",
-  "starlette==0.41.3",
-  "string-color==1.2.3",
-  "types-pytz==2023.3.1.1",
-  "verboselogs==1.7",
+  "bs4==0.0.2",               # https://pypi.org/project/bs4
+  "click==8.1.8",             # https://pypi.org/project/click
+  "coloredlogs==15.0.1",      # https://pypi.org/project/coloredlogs
+  "holidays==0.64",           # https://pypi.org/project/holidays
+  "httpx==0.28.1",            # https://pypi.org/project/httpx
+  "pydantic==2.10.5",         # https://pypi.org/project/pydantic
+  "pydantic-settings==2.7.1", # https://pypi.org/project/pydantic-settings
+  "pytz==2024.2",             # https://pypi.org/project/pytz
+  "starlette==0.45.2",        # https://pypi.org/project/starlette
+  "string-color==1.2.3",      # https://pypi.org/project/string-color
+  "verboselogs==1.7",         # https://pypi.org/project/verboselogs
 ]
 
-[project.optional-dependencies]
+[dependency-groups]
 dev = [
-  "mypy==1.7.0",
-  "pre-commit==3.5.0",
-  "pylint==3.0.2",
-  "pytest-aiofiles==0.2.0",
-  "pytest-aiohttp==1.0.5",
-  "pytest-asyncio==0.24.0",
-  "pytest-cov==6.0.0",
-  "pytest-env==1.1.5",
-  "pytest-timeout==2.3.1",
-  "pytest==8.3.3",
-  "ruff>=0.8.0",
-  "testfixtures==7.2.2",
-  "types-aiofiles==23.2.0.0",
-  "types-cachetools==5.3.0.7",
-  "types-setuptools>=75.1.0.20240917",
+  "mypy>=1.14.1",                     # https://pypi.org/project/mypy
+  "pre-commit>=4.0.1",                # https://pypi.org/project/pre-commit
+  "pycountry>=24.6.1",                # https://pypi.org/project/pycountry
+  "pylint>=3.3.3",                    # https://pypi.org/project/pylint
+  "pytest-aiofiles>=0.2.0",           # https://pypi.org/project/pytest-aiofiles
+  "pytest-aiohttp>=1.0.5",            # https://pypi.org/project/pytest-aiohttp
+  "pytest-asyncio>=0.25.2",           # https://pypi.org/project/pytest-asyncio
+  "pytest-cov>=6.0.0",                # https://pypi.org/project/pytest-cov
+  "pytest-docker-fixtures>=1.3.19",   # https://pypi.org/project/pytest-docker-fixtures
+  "pytest-env>=1.1.5",                # https://pypi.org/project/pytest-env
+  "pytest-timeout>=2.3.1",            # https://pypi.org/project/pytest-timeout
+  "pytest>=8.3.4",                    # https://pypi.org/project/pytest
+  "ruff>=0.9.1",                      # https://pypi.org/project/ruff
+  "testfixtures>=8.3.0",              # https://pypi.org/project/testfixtures
+  "types-aiofiles>=24.1.0.20241221",  # https://pypi.org/project/types-aiofiles
+  "types-cachetools>=5.5.0.20240820", # https://pypi.org/project/types-cachetools
+  "types-mock>=5.1.0.20240425",       # https://pypi.org/project/types-mock
+  "types-pillow>=10.2.0.20240822",    # https://pypi.org/project/types-pillow
+  "types-pytz>=2024.2.0.20241221",    # https://pypi.org/project/types-pytz
 ]
 
 [project.scripts]
 vm-clockify = "vm_clockify.main:cli"
 
-# ----------------------------------- BLACK -----------------------------------
-[tool.black]
-target-version = ['py312']
-# safe = true
-line-length = 130
+[tool.hatch.build.targets.sdist]
+include = ["/vm_clockify"]
+
+[tool.hatch.version]
+source = "vcs"
 
 # ____________________________________ RUFF ____________________________________
 # https://docs.astral.sh/ruff/settings/
 [tool.ruff]
 line-length = 130
-target-version = "py312"
 fix = true
 
-lint.select = [
-  "ASYNC",   # https://docs.astral.sh/ruff/rules/#flake8-async
-  "B",       # https://docs.astral.sh/ruff/rules/#flake8-bugbear
-  "C",       # complexity
-  "C4",      # https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
-  "COM818",  # Trailing comma on bare tuple prohibited
-  "D",       # docstrings
-  "DTZ003",  # Use datetime.now(tz=) instead of datetime.utcnow()
-  "DTZ004",  # Use datetime.fromtimestamp(ts, tz=) instead of datetime.utcfromtimestamp(ts)
-  "E",       # pycodestyle
-  "F",       # pyflakes/autoflake
-  "G",       # https://docs.astral.sh/ruff/rules/#flake8-logging-format-g
-  "I",       # isort
-  "ICN001",  # import concentions; {name} should be imported as {asname}
-  "ISC",     # https://docs.astral.sh/ruff/rules/#flake8-implicit-str-concat-isc
-  "N",       # pep8-naming
-  "S307",    # No builtin eval() allowed
-  "PGH004",  # Use specific rule codes when using noqa
-  "PIE",     # https://docs.astral.sh/ruff/rules/#flake8-pie-pie
-  "PLC",     # pylint
-  "PLC0414", # Useless import alias. Import alias does not rename original package.
-  "PLE",     # pylint
-  "PLR",     # pylint
-  "PLW",     # pylint
-  "PT",      # https://docs.astral.sh/ruff/rules/#flake8-pytest-style-pt
-  "PYI",     # https://docs.astral.sh/ruff/rules/#flake8-pyi-pyi
-  "Q",       # https://docs.astral.sh/ruff/rules/#flake8-quotes-q
-  "RET",     # https://docs.astral.sh/ruff/rules/#flake8-return-ret
-  "RSE",     # https://docs.astral.sh/ruff/rules/#flake8-raise-rse
-  "RUF006",  # Store a reference to the return value of asyncio.create_task
-  "S",       # flake8-bandit
-  "SIM",     # https://docs.astral.sh/ruff/rules/#flake8-simplify-sim
-  "SLF",     # https://docs.astral.sh/ruff/rules/#flake8-self-slf
-  "SLOT",    # https://docs.astral.sh/ruff/rules/#flake8-slots-slot
-  "T100",    # Trace found: {name} used
-  "T20",     # https://docs.astral.sh/ruff/rules/#flake8-print-t20
-  "TID",     # https://docs.astral.sh/ruff/rules/#flake8-tidy-imports-tid
-  "TRY004",  # Prefer TypeError exception for invalid type
-  "B904",    # Use raise from to specify exception cause
-  "TRY203",  # Remove exception handler; error is immediately re-raised
-  "UP",      # pyupgrade
-  "W",       # pycodestyle
+[tool.ruff.format]
+# quote-style = "single"
+indent-style = "space"
+docstring-code-format = true
+
+[tool.ruff.lint]
+select = ["ALL"]
+
+extend-select = [
+  "UP", # pyupgrade
+  "D",  # pydocstyle
 ]
 
-lint.ignore = [
+ignore = [
   "G004",    # Logging statement uses f-string
   "PLR2004", # Magic value used in comparison, consider replacing 5 with a constant variable
+  "D107",    # Missing docstring in `__init__`
+  "ERA001",  # Found commented-out code
+  "FBT001",  # Boolean-typed positional argument in function definition
+  "FBT002",  # Boolean default positional argument in function definition
+  "BLE001",  # Do not catch blind exception: `Exception`
+  "TRY300",  # Consider moving this statement to an `else` block
+  "RUF012",  # Mutable class attributes should be annotated with `typing.ClassVar`
+  "TD002",   # Missing author in TODO; try: `# TODO(<author_name>): ...` or `# TODO @<author_name>: ...`
+  "TD003",   # Missing issue link on the line following this TODO
 ]
 
 [tool.ruff.lint.per-file-ignores]
+"__init__.py" = [
+  "F403", # Checks for the use of wildcard imports.
+  "F405", # Checks for names that might be undefined
+]
 "tests/**" = [
   "D100",   # Missing docstring in public module
   "D103",   # Missing docstring in public function
