@@ -1,6 +1,7 @@
 """LOG HELPER."""
 
 import logging
+import sys
 
 import coloredlogs
 import verboselogs
@@ -9,42 +10,39 @@ from vm_clockify.utils.config import settings
 
 
 class LogHelper:
-    """LOGGING."""
+    """LogHelper."""
 
-    def __init__(
-        self,
-        logging_verbose: int = settings.LOGGING_VERBOSE,
-        logging_level: str = settings.LOGGING_LEVEL,
-    ):
-        """INIT LOGGING."""
+    def __init__(self) -> None:
+        """Log Helper init."""
+        logger_name_size = self._clean_logs()
         # configure logger for requested verbosity
-        if logging_verbose >= 4:
-            log_format = "[%(asctime)s,%(msecs)03d] %(name)s[%(process)d] \
-                          {%(lineno)-6d: (%(funcName)-30s)} %(levelname)-7s - %(message)s"
-        elif logging_verbose >= 3:
-            log_format = "[%(filename)-18s/%(module)-15s - %(lineno)-6d: (%(funcName)-30s)]:: %(levelname)-7s - %(message)s"
-        elif logging_verbose >= 2 or logging_verbose >= 1:
-            log_format = "%(levelname)-7s - %(message)s"
-        elif logging_verbose >= 0 or logging_verbose < 0:
-            log_format = "%(message)s"
-        else:
-            log_format = "%(message)s"
+        log_format: str = "%(message)s"
+        if settings.LOGGING_VERBOSE == 2:
+            log_format = f"[%(asctime)s] %(levelname)-5s :: %(name)-{logger_name_size}s - %(message)s"
+        elif settings.LOGGING_VERBOSE == 1:
+            log_format = "[%(asctime)s] - %(message)s"
+
+        self._clean_logs()
+        root_logger = logging.getLogger("root")
+        root_logger.setLevel(logging.getLevelName(settings.LOGGING_LEVEL))
 
         # create a log object from verboselogs
         verboselogs.install()
+        # add colored logs
+        coloredlogs.install(
+            level=logging.getLevelName(settings.LOGGING_LEVEL),
+            fmt=log_format,
+            logger=root_logger,
+            stream=sys.stdout,
+        )
 
-        for logger_name in [logging.getLogger()] + [logging.getLogger(name) for name in logging.root.manager.loggerDict]:
+    def _clean_logs(self) -> int:
+        logger_name_size = 4
+        for logger_name in [logging.getLogger()] + [logging.getLogger(name) for name in logging.getLogger().manager.loggerDict]:
             for handler in logger_name.handlers:
                 logger_name.removeHandler(handler)
-            # define log level default
-            logger_name.setLevel(logging.getLevelName(logging_level))
-            # add colered logs
-            coloredlogs.install(
-                level=logging.getLevelName(logging_level),
-                fmt=log_format,
-                logger=logger_name,
-            )
-
-            if logging_level == "INFO" and logger_name.name.startswith("httpx"):
+            if settings.LOGGING_LEVEL == "INFO" and logger_name.name.startswith("httpx"):
                 logger_name.setLevel(logging.WARNING)
-                logger_name.setLevel(logging.WARNING)
+            if (logger_name_size := len(logger_name.name)) >= logger_name_size:
+                pass
+        return logger_name_size + 1
